requirements.txt
retell-sdk
python-dotenv
fastapi
uvicorn
google-generativeai
